+++ DatadogSessionReplay/Tests/SessionReplayTests.swift
@@ -126,6 +126,22 @@ class SessionReplayTests: XCTestCase {
         XCTAssertEqual(sr.recordingCoordinator.sampler.samplingRate, random)
     }
 
+    func testWhenEnabled_itUpdateCoreContext() throws {
+        let core = PassthroughCoreMock()
+        let random: SampleRate = .mockRandom(min: 0, max: 100)
+        let startRecordingImmediately: Bool = .mockRandom()
+        config.replaySampleRate = random
+        config.startRecordingImmediately = startRecordingImmediately
+
+        // When
+        SessionReplay.enable(with: config, in: core)
+
+        // Then
+        let config = try XCTUnwrap(core.context.additionalContext(ofType: SessionReplayCoreContext.Configuration.self))
+        XCTAssertEqual(config.sampleRate, random)
+        XCTAssertEqual(config.startRecordingManually, !startRecordingImmediately)
+    }
+
     func testWhenEnabledWithCustomEndpoint() throws {
         let random: URL = .mockRandom()
         config.customEndpoint = random